+++ src/valheim_save_tools_py/__init__.py
@@ -7,13 +7,16 @@ from .exceptions import (
     JavaNotFoundError,
     CommandExecutionError
 )
+from .valheimItemReader import (parse_items_from_base64, ValheimItemReader)
 
-__version__ = "0.1.0"
+__version__ = "0.3.0"
 __all__ = [
     "ValheimSaveTools",
     "SaveFileProcessor",
     "ValheimSaveToolsError",
     "JarNotFoundError",
     "JavaNotFoundError",
-    "CommandExecutionError"
+    "CommandExecutionError",
+    "parse_items_from_base64",
+    "ValheimItemReader",
 ]